test(agent): isolate snapshot failure policy coverage
